finish refactor for shape-divider block. relates to #504.
