+++ src/blocks/shape-divider/index.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-import classnames from 'classnames';
 import includes from 'lodash/includes';
 
 /**
@@ -9,17 +8,18 @@ import includes from 'lodash/includes';
  */
 import './styles/editor.scss';
 import './styles/style.scss';
-import edit from './components/edit';
-import icons from './components/icons';
-import dividers from './components/dividers';
+import edit from './edit';
+import icons from './icons';
+import dividers from './dividers';
+import save from './save';
+import transforms from './transforms';
+import deprecated from './deprecated';
 import metadata from './block.json';
 
 /**
  * WordPress dependencies
  */
 const { __, _x } = wp.i18n;
-const { createBlock } = wp.blocks;
-const { getColorClassName } = wp.blockEditor;
 
 /**
  * Return the appropriate SVG for the block style.
@@ -60,19 +60,27 @@ export function getDividerFromStyle( className ) {
 /**
  * Block constants
  */
-const { name } = metadata;
+const { attributes, name } = metadata;
+
+const title = __( 'Shape Divider' );
 
 const icon = icons.shapeDivider;
 
 const settings = {
-	title: __( 'Shape Divider' ),
+	title,
+
 	description: __( 'Add a shape divider to visually distinquish page sections.' ),
+
 	keywords: [ __( 'hr' ), __( 'separator' ), __( 'svg' ) ],
-	attributes: metadata.attributes,
+
+	attributes,
+
 	supports: {
 		align: [ 'wide', 'full' ],
 		coBlocksSpacing: true,
+
 	},
+
 	styles: [
 		{ name: 'wavy', label: _x( 'Wavy', 'block style' ), isDefault: true },
 		{ name: 'hills', label: _x( 'Hills', 'block style' ) },
@@ -83,268 +91,15 @@ const settings = {
 		{ name: 'triangle', label: _x( 'Triangle', 'block style' ) },
 		{ name: 'pointed', label: _x( 'Pointed', 'block style' ) },
 	],
-	transforms: {
-		from: [
-			// Default.
-			...[ ':divider', ':top-divider' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform() {
-					return createBlock( `coblocks/${ name }` );
-				},
-			} ) ),
-			{
-				type: 'prefix',
-				prefix: ':bottom-divider',
-				transform: function() {
-					return createBlock( `coblocks/${ name }`, {
-						verticalFlip: true,
-					} );
-				},
-			},
-			// Waves.
-			...[ ':waves-divider', ':waves-angled-divider' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-waves',
-					} );
-				},
-			} ) ),
-			{
-				type: 'prefix',
-				prefix: ':bottom-waves-divider',
-				transform: function() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-waves',
-						verticalFlip: true,
-					} );
-				},
-			},
-			// Sloped.
-			...[ ':sloped-divider', ':top-sloped-divider' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-sloped',
-					} );
-				},
-			} ) ),
-			{
-				type: 'prefix',
-				prefix: ':bottom-sloped-divider',
-				transform: function() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-sloped',
-						verticalFlip: true,
-					} );
-				},
-			},
-			// Sloped.
-			...[ ':rounded-divider', ':top-rounded-divider' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-rounded',
-					} );
-				},
-			} ) ),
-			{
-				type: 'prefix',
-				prefix: ':bottom-rounded-divider',
-				transform: function() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-rounded',
-						verticalFlip: true,
-					} );
-				},
-			},
-			// Angled.
-			...[ ':angled-divider', ':top-angled-divider' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-angled',
-					} );
-				},
-			} ) ),
-			{
-				type: 'prefix',
-				prefix: ':bottom-angled-divider',
-				transform: function() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-angled',
-						verticalFlip: true,
-					} );
-				},
-			},
-			// Triangle.
-			...[ ':triangle-divider', ':top-triangle-divider' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-triangle',
-					} );
-				},
-			} ) ),
-			{
-				type: 'prefix',
-				prefix: ':bottom-triangle-divider',
-				transform: function() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-triangle',
-						verticalFlip: true,
-					} );
-				},
-			},
-			// Pointed.
-			...[ ':pointed-divider', ':top-pointed-divider' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-pointed',
-					} );
-				},
-			} ) ),
-			{
-				type: 'prefix',
-				prefix: ':bottom-pointed-divider',
-				transform: function() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-pointed',
-						verticalFlip: true,
-					} );
-				},
-			},
-			// Hills.
-			...[ ':hills-divider', ':top-hills-divider' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-hills',
-					} );
-				},
-			} ) ),
-			{
-				type: 'prefix',
-				prefix: ':bottom-hills-divider',
-				transform: function() {
-					return createBlock( `coblocks/${ name }`, {
-						className: 'is-style-hills',
-						verticalFlip: true,
-					} );
-				},
-			},
-			{
-				type: 'block',
-				blocks: [ 'core/spacer' ],
-				transform: ( { height } ) => createBlock( `coblocks/${ name }`, {
-					height: height,
-				} ),
-			},
-			{
-				type: 'block',
-				blocks: [ 'coblocks/dynamic-separator' ],
-				transform: ( { height } ) => createBlock( `coblocks/${ name }`, {
-					height: height,
-				} ),
-			},
-			{
-				type: 'block',
-				blocks: [ 'core/separator' ],
-				transform: () => createBlock( `coblocks/${ name }` ),
-			},
-		],
-	},
+
+	transforms,
+
 	edit,
-	save( { attributes, className } ) {
-		const {
-			backgroundColor,
-			backgroundHeight,
-			coblocks,
-			color,
-			customBackgroundColor,
-			customColor,
-			horizontalFlip,
-			shapeHeight,
-			verticalFlip,
-		} = attributes;
 
-		const shapeClass = getColorClassName( 'color', color );
-		const backgroundClass = getColorClassName( 'background-color', backgroundColor );
+	save,
 
-		const classes = classnames(
-			className, {
-				[ `coblocks-shape-divider-${ coblocks.id }` ]: coblocks && ( typeof coblocks.id !== 'undefined' ),
-				'is-vertically-flipped': verticalFlip,
-				'is-horizontally-flipped': horizontalFlip,
-				[ shapeClass ]: shapeClass,
-				[ backgroundClass ]: backgroundClass,
-			} );
-
-		const styles = {
-			backgroundColor: backgroundClass ? undefined : customBackgroundColor,
-			color: shapeClass ? undefined : customColor,
-		};
-
-		return (
-			<div className={ classes } style={ styles } aria-hidden="true">
-				<div className="wp-block-coblocks-shape-divider__svg-wrapper" style={ { minHeight: shapeHeight } }>
-					{ getDividerFromStyle( attributes.className ) }
-				</div>
-				<div className="wp-block-coblocks-shape-divider__alt-wrapper" style={ { minHeight: backgroundHeight } }></div>
-			</div>
-		);
-	},
-	deprecated: [
-		{
-			save( { attributes, className } ) {
-				const {
-					backgroundColor,
-					backgroundHeight,
-					color,
-					customBackgroundColor,
-					customColor,
-					horizontalFlip,
-					shapeHeight,
-					verticalFlip,
-				} = attributes;
-
-				const shapeClass = getColorClassName( 'color', color );
-				const backgroundClass = getColorClassName( 'background-color', backgroundColor );
-
-				const classes = classnames(
-					className, {
-						'is-vertically-flipped': verticalFlip,
-						'is-horizontally-flipped': horizontalFlip,
-						[ shapeClass ]: shapeClass,
-						[ backgroundClass ]: backgroundClass,
-					} );
-
-				const styles = {
-					backgroundColor: backgroundClass ? undefined : customBackgroundColor,
-					color: shapeClass ? undefined : customColor,
-				};
-
-				return (
-					<div className={ classes } style={ styles } aria-hidden="true">
-						<div className="wp-block-coblocks-shape-divider__svg-wrapper" style={ { height: shapeHeight } }>
-							{ getDividerFromStyle( attributes.className ) }
-						</div>
-						<div className="wp-block-coblocks-shape-divider__alt-wrapper" style={ { height: backgroundHeight } }></div>
-					</div>
-				);
-			},
-		},
-	],
+	deprecated,
 };
 
-export { name, icon, settings };
+export { name, title, icon, settings };
